services/embed-service/requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
requests==2.32.4
qdrant-client==1.14.3
pydantic==2.11.7
google-generativeai==0.5.4